refactor(browser): isolate UI warnings and remove dead support code
